% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RSAccess-package.R
\docType{package}
\name{RSAccess-package}
\alias{RSAccess-package}
\title{RSAccess: A data package to integrate remote sensing variables and job accessibility measures for the Greater Golden Horseshoe (GGH) area in  Ontario, Canada}
\references{
\url{https://github.com/dias-bruno/RSAccess}
}
\author{
Bruno Dos Santos, McMaster University, \email{dossanb@mcmaster.ca}
}
